// src/types/chat.ts
// src/types/chat.ts

export type MessageRole = "user" | "assistant";

export interface Message {
    id: string;
    role: MessageRole;
    content: string;
    timestamp: Date;
}

export interface Conversation{
    id:string ;
    title: string;
    messages: Message[];
    createdAt: Date;
    updatedAt: Date;
}